+++ esm/tag-classes/comment-tag.js
@@ -1,4 +1,4 @@
-import Signal from '../lib/signal.js';
+import Signal, { effect } from '../lib/signal.js';
 
 export default class CommentTag {
   constructor(text) {
@@ -19,13 +19,20 @@ export default class CommentTag {
     if (typeof document === 'undefined') {
       throw new Error('toElement only supported in browser');
     }
-    if (!(this.text instanceof Signal)) {
-      return document.createComment(this.text);
+    if (this.text instanceof Signal) {
+      let comment = null;
+      const sig = this.text;
+      effect(() => {
+        const val = sig.get();
+        if (comment === null) {
+          comment = document.createComment(val);
+        } else {
+          comment.nodeValue = val;
+        }
+      });
+      return comment;
     }
-    const comment = document.createComment(this.text.get());
-    this.text.subscribe(val => {
-      comment.nodeValue = val;
-    });
-    return comment;
+
+    return document.createComment(this.text);
   }
 }